requirements.txt
django
pandas
mysqlclient
openpyxl
djangorestframework
django-cors-headers